+++ src/scenario/tests/mod.rs
@@ -2,3 +2,4 @@ mod basic;
 mod gates;
 mod run_config;
 mod setup;
+mod target;